requirements.txt
Flask==1.1.2
google-cloud-datastore
google-api-python-client
pika
kazoo
